project/server/requirements.txt
Flask
Flask-RESTful
Flask-JWT-Extended
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
